Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    int g;
-    scanf("%d",&g);
-    printf("Area of circle: %d",3.14*r*r);
+    float g;
+    scanf("%f",&g);
+    printf("Area of circle: %.2f",3.14*r*r);
     return 0;
     }
